[refactor] Such that index.js works properly renders correct component, pulled out quiz component
