+++ cohack/src/tests/App.test.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { render } from '@testing-library/react';
-import App from './App';
+import App from '../App';
 
 test('renders learn react link', () => {
   const { getByText } = render(<App />);
